Move ClientNetworking.{h,cpp} into client directory
