+++ UI/ChatWnd.cpp
@@ -3,10 +3,10 @@
 #include "CUIControls.h"
 #include "PlayerListWnd.h"
 #include "../client/human/HumanClientApp.h"
+#include "../client/ClientNetworking.h"
 #include "../Empire/Empire.h"
 #include "../Empire/EmpireManager.h"
 #include "../network/Message.h"
-#include "../network/ClientNetworking.h"
 #include "../universe/Universe.h"
 #include "../universe/Ship.h"
 #include "../universe/ShipDesign.h"
